Add function to script.js
Added alert at the start of the game to inform the user about the game
is best of 5.
Added feature to prevent other buttons being clicked while on timeout.
Added feature to ask if the player want a rematch/replay when the game
is over.
Added feature to ends round on 5.